# docs/api-specification.yaml
openapi: 3.0.3
info:
  title: Nomikai Restaurant Finder API
  description: API for finding restaurants suitable for nomikai (drinking parties)
  version: 1.0.0
  contact:
    name: Claude Code
    email: support@example.com
  license:
    name: MIT

servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.nomikai-finder.com
    description: Production server

paths:
  /api/health:
    get:
      summary: Health check endpoint
      tags:
        - Health
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
        '503':
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/health/readiness:
    get:
      summary: Readiness check endpoint
      tags:
        - Health
      responses:
        '200':
          description: Service is ready
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '503':
          description: Service not ready
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/health/liveness:
    get:
      summary: Liveness check endpoint
      tags:
        - Health
      responses:
        '200':
          description: Service is alive
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  /api/auth/register:
    post:
      summary: User registration
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/auth/login:
    post:
      summary: User login
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/restaurants/search:
    get:
      summary: Search restaurants
      tags:
        - Restaurants
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: genre
          in: query
          description: Restaurant genre filter
          schema:
            type: string
            example: "Japanese"
        - name: location
          in: query
          description: Location filter
          schema:
            type: string
            example: "Shibuya"
        - name: priceRange
          in: query
          description: Price range filter
          schema:
            type: string
            enum: [low, medium, high]
        - name: limit
          in: query
          description: Number of results to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Number of results to skip
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestaurantSearchResponse'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'

  /api/restaurants/{id}:
    get:
      summary: Get restaurant by ID
      tags:
        - Restaurants
      parameters:
        - name: id
          in: path
          required: true
          description: Restaurant ID
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Restaurant details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestaurantDetailResponse'
        '404':
          description: Restaurant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/restaurants/{id}/evaluate:
    post:
      summary: Evaluate restaurant
      tags:
        - Restaurants
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: id
          in: path
          required: true
          description: Restaurant ID
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Evaluation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluationResponse'
        '404':
          description: Restaurant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
        timestamp:
          type: string
          format: date-time

    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: "error"
        message:
          type: string
        timestamp:
          type: string
          format: date-time

    ValidationErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: "error"
        message:
          type: string
          example: "Validation failed"
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string

    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        data:
          type: object
          properties:
            status:
              type: string
              example: "healthy"
            timestamp:
              type: string
              format: date-time
            service:
              type: string
              example: "nomikai-restaurant-finder"
            version:
              type: string
              example: "1.0.0"
            checks:
              type: object
              properties:
                database:
                  type: string
                  example: "healthy"
                memory:
                  type: object
                  properties:
                    used:
                      type: number
                    total:
                      type: number
                    unit:
                      type: string
                      example: "MB"
                uptime:
                  type: number

    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          minLength: 8
          pattern: "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)"
          example: "SecurePass123"

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          minLength: 6
          example: "SecurePass123"

    AuthResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
        data:
          type: object
          properties:
            user:
              type: object
              properties:
                id:
                  type: integer
                email:
                  type: string
                  format: email
            token:
              type: string

    Restaurant:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        genre:
          type: string
        location:
          type: string
        price_range:
          type: string
          enum: [low, medium, high]
        created_at:
          type: string
          format: date-time
        avgRating:
          type: number
          format: float
          minimum: 0
          maximum: 5
        totalReviews:
          type: integer
          minimum: 0

    RestaurantSearchResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
        data:
          type: object
          properties:
            restaurants:
              type: array
              items:
                $ref: '#/components/schemas/Restaurant'
            pagination:
              type: object
              properties:
                limit:
                  type: integer
                offset:
                  type: integer
                total:
                  type: integer

    RestaurantDetailResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
        data:
          $ref: '#/components/schemas/Restaurant'

    EvaluationResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
        data:
          type: object
          properties:
            restaurantId:
              type: integer
            evaluationScore:
              type: number
              format: float